Add cancel and activate event functionalities
Two new methods for canceling and activating events are added to the events management API. The endpoint paths "/:eventId/cancel" and "/:eventId/makeActive" have been added to "event_controller.go". Corresponding service methods, "CancelEvent" and "MakeEventActive" are added to the "event_service.go" file along with the functioning code to update the status of the event. Updates to entities regarding cancellation and activation status are also reflected in the schema structure in the "event.go" file.