Add media management and attendee assignment to ticket options
Expanded functionality in the ticket controller by implementing methods to add and remove media from ticket options and to fetch a list of event attendees. Also added a feature to assign a ticket to an attendee. These updates will facilitate handling of event media and allow for easier assignment and tracking of event attendees.